+++ cli/types.ts
@@ -1,0 +1,5 @@
+import { KeypressEvent } from '@inquirer/prompts';
+
+export interface Key extends KeypressEvent {
+    sequence: string;
+}
