+++ cli/utils.ts
@@ -1,22 +1,25 @@
+import { KeypressEvent } from '@inquirer/prompts';
+
 const PARAM_REGEX = /\{\w+\}/g;
 
 // Check if a given string contains any params
-const hasParams = (s) => {
+const hasParams = (s: string) => {
     const params = s.match(PARAM_REGEX);
     if (!params) return false;
     return true;
 };
 
 // Return the number of params in a given string
-const numParams = (s) => {
+const numParams = (s: string) => {
     const params = s.match(PARAM_REGEX);
     if (!params) return 0;
     return params.length;
 };
 
-const getPathParams = (s) => {
-    const matches = s.match(PARAM_REGEX);
-    return matches.map((match) => {
+// Return a list of params in a given path
+const getPathParams = (s: string) => {
+    const matches = s.match(PARAM_REGEX) || [];
+    return matches.map((match: string) => {
         return match.slice(1, match.length - 1);
     });
 };
@@ -24,12 +27,15 @@ const getPathParams = (s) => {
 /**
  * Takes an array of API paths and returns a new array containing only paths
  * which use the given request type.
- * @param {Array<Array<string, object>>} paths
+ * @param {Array<Array<T>>} paths
  * @param {'get' | 'post' | 'put' | 'delete'} reqType
- * @returns {Array<Array<string, object>>}
+ * @returns {Array<Array<T>>}
  */
-const filterPathsByReqType = (paths, reqType) => {
-    const filteredArray = [];
+const filterPathsByReqType = <T>(
+    paths: Array<Array<T>>,
+    reqType: 'get' | 'post' | 'put' | 'delete'
+): Array<Array<T>> => {
+    const filteredArray: Array<Array<T>> = [];
     for (let i = 0; i < paths.length; i++) {
         if (paths[i][1][reqType]) filteredArray.push(paths[i]);
     }
@@ -41,7 +47,7 @@ const filterPathsByReqType = (paths, reqType) => {
  * @param {object} key ex: { sequence: 'a', name: 'a', ctrl: false, meta: false, shift: false }
  * @returns {boolean}
  */
-const isAlphaNumeric = (key) => {
+const isAlphaNumeric = (key: KeypressEvent): boolean => {
     if (key.name && key.name.length === 1 && key.name.toLowerCase().match(/[a-z]|[0-9]/i)) {
         return true;
     }
@@ -55,7 +61,7 @@ const isAlphaNumeric = (key) => {
  * @param {string} search
  * @returns {string}
  */
-const addSpaceToSearchBar = (search) => {
+const addSpaceToSearchBar = (search: string): string => {
     const diff = 50 - (search.length || 0);
     let spaces = '';
     for (let i = 0; i < diff; i++) {
@@ -70,7 +76,10 @@ const addSpaceToSearchBar = (search) => {
  * @param {Array<{ param: string, value: string, path: boolean }>} params
  * @returns {Promise<object>}
  */
-const fetchTMDB = async (path, params) => {
+const fetchTMDB = async (
+    path: string,
+    params: Array<{ param: string; value: string; path: boolean }>
+): Promise<object> => {
     const BASE_PATH = 'https://api.themoviedb.org';
     // eslint-disable-next-line no-undef
     const API_KEY = '?api_key=' + process.env.VITE_MOVIEDB_KEY;
