+++ package.json
@@ -25,7 +25,8 @@
     "db-start-env": "./scripts/db-start.sh",
     "db-stop": "npx supabase stop",
     "db-reset": "npx supabase db reset",
-    "patch": "patch-package"
+    "patch": "patch-package",
+    "cli": "tsx cli/index.ts"
   },
   "dependencies": {
     "@emotion/react": "^11.11.1",
@@ -78,7 +79,12 @@
     "storybook-addon-react-router-v6": "^2.0.10",
     "supabase": "^1.110.1",
     "tailwindcss": "^3.3.5",
+<<<<<<< HEAD
     "typescript": "^5.3.3",
+=======
+    "tsx": "^4.7.0",
+    "typescript": "^5.1.6",
+>>>>>>> 6592215 (convert to typescript)
     "vite": "^4.4.12",
     "vitest": "^1.1.3"
   }
